docs: 📝 Update README.md with info about abstract class relations
Added a warning and solution to the README about using relations with abstract classes.

When using an abstract class for a relation property, it must be declared `open` and have a default value. Otherwise, a Kotlin inheritance error will be thrown. This is now documented in the README with a warning and a tip.