docs: ✍️ Add `CancellationException` to `@Throws` annotations
This commit updates the KDoc for several suspend functions to explicitly declare that they can throw a `CancellationException`.

This improves the correctness of the API documentation by making it clear to developers that these coroutine-based network operations are cancellable.

Specific changes include adding `CancellationException` to the `@Throws` annotation for the following methods:

- `Scope.find()`
- `Scope.first()`
- `Scope.last()`
- `Resource.save()`
- `Resource.destroy()`